12/03/2023

+++ AgroTech/frontend/pages/login/js/login.js
@@ -1,5 +1,5 @@
-const inpUser = document.querySelector('#user')
-const inpSenha = document.querySelector('#senha')
+var inpUser = document.querySelector('#user')
+var inpSenha = document.querySelector('#senha')
 
 var uriCard_Usuarios = 'http://localhost:3000/usuarios'
 
@@ -18,45 +18,92 @@ fetch(uriCard_Usuarios, options)
 
 
 function logar() {
-    let data = {
-        "email": inpUser.value,
-        "senha": inpSenha.value
+
+    document.querySelector('.e1').classList.add('model')
+    document.querySelector('.e2').classList.add('model')
+    document.querySelector('.e3').classList.add('model')
+
+
+    var existeEmail = false;
+    var existeSenha = false;
+    var erroVazio = false;
+
+
+console.log(inpUser.value, inpSenha.value);
+
+    if (inpUser.value == "" || inpSenha.value == "") {
+        document.querySelector('.e1').classList.remove('model')
+        erroVazio = true
     }
-    fetch("http://localhost:3000/usuarios/login", {
-        "method": "POST",
-        "headers": {
-            "Content-Type": "application/json"
-        },
-        "body": JSON.stringify(data)
-    })
-        .then(res => { return res.json() })
-        .then(data => {
 
-            if (data.erro === undefined) {
-                console.log(data)
-                localStorage.setItem("info", JSON.stringify({ "id_user": data.uid, "nome": data.uname, "token": data.token }));
+    if (erroVazio == false) {
+
+        usuarios.forEach(u => {
 
 
-                usuarios.forEach(u => {
-                    if(u.id == data.uid) {
+            if (u.email == inpUser.value) {
+                existeEmail = true;
+            }
+
+        })
+
+        if (existeEmail == false) {
+            document.querySelector('.e2').classList.remove('model')
 
 
-                        if(u.tipo == "usuario") {
-                            window.location.href = '../AreaComum/areaComum.html'
-                            console.log('useres');
+        }
 
-                        }
+            if (existeEmail == true) {
 
-                        if(u.tipo == "gerente") {
-                        console.log('geres');
-
-                             window.location.href = '../../home.html'
-
-                        }
-                    }
+                let data = {
+                    "email": inpUser.value,
+                    "senha": inpSenha.value
+                }
+                fetch("http://localhost:3000/usuarios/login", {
+                    "method": "POST",
+                    "headers": {
+                        "Content-Type": "application/json"
+                    },
+                    "body": JSON.stringify(data)
                 })
+                    .then(res => { return res.json() })
+                    .then(data => {
+
+                        if(data.erro == "Senha inválida") {
+                            document.querySelector('.e3').classList.remove('model')
+                        }
+
+                        if (data.erro === undefined) {
+                            console.log(data)
+                            localStorage.setItem("info", JSON.stringify({ "id_user": data.uid, "nome": data.uname, "token": data.token }));
+
+
+                            usuarios.forEach(u => {
+                                if (u.id == data.uid) {
+
+
+                                    if (u.tipo == "usuario") {
+                                        window.location.href = '../AreaComum/areaComum.html'
+                                        console.log('useres');
+
+                                    }
+
+                                    if (u.tipo == "gerente") {
+                                        console.log('geres');
+
+                                        window.location.href = '../../home.html'
+
+                                    }
+                                }
+                            })
+
+                        }
+                    })
+
 
             }
-        })
+
+        }
+
 }
